requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.1
asn1crypto==1.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect==1.1.0
cssselect2==0.6.0
cycler==0.11.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
django-ajax-datatable==4.4.4
django-heroku==0.3.1
django-jsignature==0.11
fonttools==4.33.3
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.2
mysqlclient==2.1.0
numpy==1.22.3
oscrypto==1.3.0
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
pyparsing==3.0.9
PyPDF3==1.0.6
pyquery==1.4.3
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.1.0
xhtml2pdf==0.2.7
